# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MyBrokeLife API",
    "description": "API for personal finance tracking including authentication, categories, transactions, budgets and reports.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Use 'Bearer <token>' for JWT authentication"
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/categories": {
      "get": {
        "tags": ["Categories", "Untested"],
        "summary": "Get all categories for the authenticated user",
        "responses": {
          "200": {
            "description": "A list of categories",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Category" }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories", "Untested"],
        "summary": "Create a new category",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CategoryInput" }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
...
    "/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get all transactions for the authenticated user",
        "parameters": [
          { "name": "from", "in": "query", "type": "string", "format": "date" },
          { "name": "to", "in": "query", "type": "string", "format": "date" },
          {
            "name": "type",
            "in": "query",
            "type": "string",
            "enum": ["expense", "income"]
          },
          { "name": "category", "in": "query", "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "A list of transactions",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Transaction" }
            }
          }
        }
      },
      "post": {
        "tags": ["Transactions"],
        "summary": "Create a new transaction",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/TransactionInput" }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
...
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "budgetId": { "type": "string" },
        "userId": { "type": "string" },
        "name": { "type": "string" },
        "type": { "type": "string", "enum": ["expense", "income"] },
        "color": { "type": "string" },
        "icon": { "type": "string" },
        "createdAt": { "type": "string" },
        "updatedAt": { "type": "string" }
      }
    },
    "CategoryInput": {
      "type": "object",
      "required": ["budgetId", "userId", "name", "type"],
      "properties": {
        "budgetId": { "type": "string" },
        "userId": { "type": "string" },
        "name": { "type": "string" },
        "type": { "type": "string", "enum": ["expense", "income"] },
        "color": { "type": "string" },
        "icon": { "type": "string" }
      }
    },
    "Transaction": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "userId": { "type": "string" },
        "categoryId": { "type": "string" },
        "amount": { "type": "number" },
        "type": { "type": "string", "enum": ["expense", "income"] },
        "date": { "type": "string" },
        "note": { "type": "string" },
        "account": { "type": "string" },
        "createdAt": { "type": "string" },
        "updatedAt": { "type": "string" }
      }
    },
    "TransactionInput": {
      "type": "object",
      "required": ["categoryId", "amount", "type", "date"],
      "properties": {
        "categoryId": { "type": "string" },
        "amount": { "type": "number" },
        "type": { "type": "string", "enum": ["expense", "income"] },
        "date": { "type": "string", "format": "date-time" },
        "note": { "type": "string" },
        "account": { "type": "string" }
      }
    },
...
  }
}